chore: duplicated origins and tracks are merged for 'cere' and 'cere_dev' environment
