+++ runtime/cere-dev/Cargo.toml
@@ -104,6 +104,7 @@ pallet-ddc-payouts = { workspace = true }
 pallet-ddc-staking = { workspace = true }
 pallet-erc20 = { workspace = true }
 pallet-erc721 = { workspace = true }
+pallet-origins = { workspace = true }
 
 [build-dependencies]
 substrate-wasm-builder = { workspace = true, default-features = true }
@@ -191,6 +192,7 @@ std = [
   "pallet-whitelist/std",
   "pallet-ddc-clusters-gov/std",
   "sp-arithmetic/std",
+  "pallet-origins/std",
 ]
 runtime-benchmarks = [
   "frame-benchmarking/runtime-benchmarks",
