+++ runtime/cere/src/governance/tracks.rs
@@ -1,279 +1,9 @@
-// Copyright 2023 Parity Technologies (UK) Ltd.
-// This file is part of Polkadot.
-
-// Polkadot is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// Polkadot is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with Polkadot. If not, see <http://www.gnu.org/licenses/>.
-
 //! Track configurations for governance.
 
+use cere_runtime_common::constants::tracks::TRACKS_DATA;
+
 use super::*;
-
-const fn percent(x: i32) -> sp_arithmetic::FixedI64 {
-	sp_arithmetic::FixedI64::from_rational(x as u128, 100)
-}
-use pallet_referenda::Curve;
-const APP_ROOT: Curve = Curve::make_reciprocal(4, 28, percent(80), percent(50), percent(100));
-const SUP_ROOT: Curve = Curve::make_linear(28, 28, percent(0), percent(50));
-const APP_STAKING_ADMIN: Curve = Curve::make_linear(17, 28, percent(50), percent(100));
-const SUP_STAKING_ADMIN: Curve =
-	Curve::make_reciprocal(12, 28, percent(1), percent(0), percent(50));
-const APP_TREASURER: Curve = Curve::make_reciprocal(4, 28, percent(80), percent(50), percent(100));
-const SUP_TREASURER: Curve = Curve::make_linear(28, 28, percent(0), percent(50));
-const APP_FELLOWSHIP_ADMIN: Curve = Curve::make_linear(17, 28, percent(50), percent(100));
-const SUP_FELLOWSHIP_ADMIN: Curve =
-	Curve::make_reciprocal(12, 28, percent(1), percent(0), percent(50));
-const APP_GENERAL_ADMIN: Curve =
-	Curve::make_reciprocal(4, 28, percent(80), percent(50), percent(100));
-const SUP_GENERAL_ADMIN: Curve =
-	Curve::make_reciprocal(7, 28, percent(10), percent(0), percent(50));
-const APP_REFERENDUM_CANCELLER: Curve = Curve::make_linear(17, 28, percent(50), percent(100));
-const SUP_REFERENDUM_CANCELLER: Curve =
-	Curve::make_reciprocal(12, 28, percent(1), percent(0), percent(50));
-const APP_REFERENDUM_KILLER: Curve = Curve::make_linear(17, 28, percent(50), percent(100));
-const SUP_REFERENDUM_KILLER: Curve =
-	Curve::make_reciprocal(12, 28, percent(1), percent(0), percent(50));
-const APP_SMALL_TIPPER: Curve = Curve::make_linear(10, 28, percent(50), percent(100));
-const SUP_SMALL_TIPPER: Curve = Curve::make_reciprocal(1, 28, percent(4), percent(0), percent(50));
-const APP_BIG_TIPPER: Curve = Curve::make_linear(10, 28, percent(50), percent(100));
-const SUP_BIG_TIPPER: Curve = Curve::make_reciprocal(8, 28, percent(1), percent(0), percent(50));
-const APP_SMALL_SPENDER: Curve = Curve::make_linear(17, 28, percent(50), percent(100));
-const SUP_SMALL_SPENDER: Curve =
-	Curve::make_reciprocal(12, 28, percent(1), percent(0), percent(50));
-const APP_MEDIUM_SPENDER: Curve = Curve::make_linear(23, 28, percent(50), percent(100));
-const SUP_MEDIUM_SPENDER: Curve =
-	Curve::make_reciprocal(16, 28, percent(1), percent(0), percent(50));
-const APP_BIG_SPENDER: Curve = Curve::make_linear(28, 28, percent(50), percent(100));
-const SUP_BIG_SPENDER: Curve = Curve::make_reciprocal(20, 28, percent(1), percent(0), percent(50));
-const APP_WHITELISTED_CALLER: Curve =
-	Curve::make_reciprocal(16, 28 * 24, percent(96), percent(50), percent(100));
-const SUP_WHITELISTED_CALLER: Curve =
-	Curve::make_reciprocal(1, 28, percent(20), percent(5), percent(50));
-
-const TRACKS_DATA: [(u16, pallet_referenda::TrackInfo<Balance, BlockNumber>); 15] = [
-	(
-		0,
-		pallet_referenda::TrackInfo {
-			name: "root",
-			max_deciding: 1,
-			decision_deposit: 100 * GRAND,
-			prepare_period: 2 * HOURS,
-			decision_period: 28 * DAYS,
-			confirm_period: 24 * HOURS,
-			min_enactment_period: 24 * HOURS,
-			min_approval: APP_ROOT,
-			min_support: SUP_ROOT,
-		},
-	),
-	(
-		1,
-		pallet_referenda::TrackInfo {
-			name: "whitelisted_caller",
-			max_deciding: 100,
-			decision_deposit: 10 * GRAND,
-			prepare_period: 30 * MINUTES,
-			decision_period: 28 * DAYS,
-			confirm_period: 10 * MINUTES,
-			min_enactment_period: 10 * MINUTES,
-			min_approval: APP_WHITELISTED_CALLER,
-			min_support: SUP_WHITELISTED_CALLER,
-		},
-	),
-	(
-		10,
-		pallet_referenda::TrackInfo {
-			name: "staking_admin",
-			max_deciding: 10,
-			decision_deposit: 5 * GRAND,
-			prepare_period: 2 * HOURS,
-			decision_period: 28 * DAYS,
-			confirm_period: 3 * HOURS,
-			min_enactment_period: 10 * MINUTES,
-			min_approval: APP_STAKING_ADMIN,
-			min_support: SUP_STAKING_ADMIN,
-		},
-	),
-	(
-		11,
-		pallet_referenda::TrackInfo {
-			name: "treasurer",
-			max_deciding: 10,
-			decision_deposit: GRAND,
-			prepare_period: 2 * HOURS,
-			decision_period: 28 * DAYS,
-			confirm_period: 3 * HOURS,
-			min_enactment_period: 24 * HOURS,
-			min_approval: APP_TREASURER,
-			min_support: SUP_TREASURER,
-		},
-	),
-	(
-		13,
-		pallet_referenda::TrackInfo {
-			name: "fellowship_admin",
-			max_deciding: 10,
-			decision_deposit: 5 * GRAND,
-			prepare_period: 2 * HOURS,
-			decision_period: 28 * DAYS,
-			confirm_period: 3 * HOURS,
-			min_enactment_period: 10 * MINUTES,
-			min_approval: APP_FELLOWSHIP_ADMIN,
-			min_support: SUP_FELLOWSHIP_ADMIN,
-		},
-	),
-	(
-		14,
-		pallet_referenda::TrackInfo {
-			name: "general_admin",
-			max_deciding: 10,
-			decision_deposit: 5 * GRAND,
-			prepare_period: 2 * HOURS,
-			decision_period: 28 * DAYS,
-			confirm_period: 3 * HOURS,
-			min_enactment_period: 10 * MINUTES,
-			min_approval: APP_GENERAL_ADMIN,
-			min_support: SUP_GENERAL_ADMIN,
-		},
-	),
-	(
-		20,
-		pallet_referenda::TrackInfo {
-			name: "referendum_canceller",
-			max_deciding: 1_000,
-			decision_deposit: 10 * GRAND,
-			prepare_period: 2 * HOURS,
-			decision_period: 7 * DAYS,
-			confirm_period: 3 * HOURS,
-			min_enactment_period: 10 * MINUTES,
-			min_approval: APP_REFERENDUM_CANCELLER,
-			min_support: SUP_REFERENDUM_CANCELLER,
-		},
-	),
-	(
-		21,
-		pallet_referenda::TrackInfo {
-			name: "referendum_killer",
-			max_deciding: 1_000,
-			decision_deposit: 50 * GRAND,
-			prepare_period: 2 * HOURS,
-			decision_period: 28 * DAYS,
-			confirm_period: 3 * HOURS,
-			min_enactment_period: 10 * MINUTES,
-			min_approval: APP_REFERENDUM_KILLER,
-			min_support: SUP_REFERENDUM_KILLER,
-		},
-	),
-	(
-		30,
-		pallet_referenda::TrackInfo {
-			name: "small_tipper",
-			max_deciding: 200,
-			decision_deposit: DOLLARS,
-			prepare_period: MINUTES,
-			decision_period: 7 * DAYS,
-			confirm_period: 10 * MINUTES,
-			min_enactment_period: MINUTES,
-			min_approval: APP_SMALL_TIPPER,
-			min_support: SUP_SMALL_TIPPER,
-		},
-	),
-	(
-		31,
-		pallet_referenda::TrackInfo {
-			name: "big_tipper",
-			max_deciding: 100,
-			decision_deposit: 10 * DOLLARS,
-			prepare_period: 10 * MINUTES,
-			decision_period: 7 * DAYS,
-			confirm_period: HOURS,
-			min_enactment_period: 10 * MINUTES,
-			min_approval: APP_BIG_TIPPER,
-			min_support: SUP_BIG_TIPPER,
-		},
-	),
-	(
-		32,
-		pallet_referenda::TrackInfo {
-			name: "small_spender",
-			max_deciding: 50,
-			decision_deposit: 100 * DOLLARS,
-			prepare_period: 4 * HOURS,
-			decision_period: 28 * DAYS,
-			confirm_period: 12 * HOURS,
-			min_enactment_period: 24 * HOURS,
-			min_approval: APP_SMALL_SPENDER,
-			min_support: SUP_SMALL_SPENDER,
-		},
-	),
-	(
-		33,
-		pallet_referenda::TrackInfo {
-			name: "medium_spender",
-			max_deciding: 50,
-			decision_deposit: 200 * DOLLARS,
-			prepare_period: 4 * HOURS,
-			decision_period: 28 * DAYS,
-			confirm_period: 24 * HOURS,
-			min_enactment_period: 24 * HOURS,
-			min_approval: APP_MEDIUM_SPENDER,
-			min_support: SUP_MEDIUM_SPENDER,
-		},
-	),
-	(
-		34,
-		pallet_referenda::TrackInfo {
-			name: "big_spender",
-			max_deciding: 50,
-			decision_deposit: 400 * DOLLARS,
-			prepare_period: 4 * HOURS,
-			decision_period: 28 * DAYS,
-			confirm_period: 48 * HOURS,
-			min_enactment_period: 24 * HOURS,
-			min_approval: APP_BIG_SPENDER,
-			min_support: SUP_BIG_SPENDER,
-		},
-	),
-	(
-		100,
-		pallet_referenda::TrackInfo {
-			name: "cluster_activator",
-			max_deciding: 50,
-			decision_deposit: 400 * DOLLARS,
-			prepare_period: 4 * HOURS,
-			decision_period: 28 * DAYS,
-			confirm_period: 48 * HOURS,
-			min_enactment_period: 24 * HOURS,
-			min_approval: APP_BIG_SPENDER,
-			min_support: SUP_BIG_SPENDER,
-		},
-	),
-	(
-		101,
-		pallet_referenda::TrackInfo {
-			name: "cluster_economics_updater",
-			max_deciding: 50,
-			decision_deposit: 400 * DOLLARS,
-			prepare_period: 4 * HOURS,
-			decision_period: 28 * DAYS,
-			confirm_period: 48 * HOURS,
-			min_enactment_period: 24 * HOURS,
-			min_approval: APP_BIG_SPENDER,
-			min_support: SUP_BIG_SPENDER,
-		},
-	),
-];
-
-pub const CLUSTER_ACTIVATOR_TRACK_ID: u16 = 100;
-pub const CLUSTER_ECONOMICS_UPDATER_TRACK_ID: u16 = 101;
+use crate::{Balance, BlockNumber};
 
 pub struct TracksInfo;
 impl pallet_referenda::TracksInfo<Balance, BlockNumber> for TracksInfo {
@@ -288,30 +18,33 @@ impl pallet_referenda::TracksInfo<Balance, BlockNumber> for TracksInfo {
 				frame_system::RawOrigin::Root => Ok(0),
 				_ => Err(()),
 			}
-		} else if let Ok(custom_origin) = origins::Origin::try_from(id.clone()) {
+		} else if let Ok(custom_origin) = pallet_origins::pallet::Origin::try_from(id.clone()) {
 			match custom_origin {
-				origins::Origin::WhitelistedCaller => Ok(1),
+				pallet_origins::pallet::Origin::WhitelistedCaller => Ok(1),
 				// General admin
-				origins::Origin::StakingAdmin => Ok(10),
-				origins::Origin::Treasurer => Ok(11),
-				origins::Origin::FellowshipAdmin => Ok(13),
-				origins::Origin::GeneralAdmin => Ok(14),
+				pallet_origins::pallet::Origin::StakingAdmin => Ok(10),
+				pallet_origins::pallet::Origin::Treasurer => Ok(11),
+				pallet_origins::pallet::Origin::FellowshipAdmin => Ok(13),
+				pallet_origins::pallet::Origin::GeneralAdmin => Ok(14),
 				// Referendum admins
-				origins::Origin::ReferendumCanceller => Ok(20),
-				origins::Origin::ReferendumKiller => Ok(21),
+				pallet_origins::pallet::Origin::ReferendumCanceller => Ok(20),
+				pallet_origins::pallet::Origin::ReferendumKiller => Ok(21),
 				// Limited treasury spenders
-				origins::Origin::SmallTipper => Ok(30),
-				origins::Origin::BigTipper => Ok(31),
-				origins::Origin::SmallSpender => Ok(32),
-				origins::Origin::MediumSpender => Ok(33),
-				origins::Origin::BigSpender => Ok(34),
+				pallet_origins::pallet::Origin::SmallTipper => Ok(30),
+				pallet_origins::pallet::Origin::BigTipper => Ok(31),
+				pallet_origins::pallet::Origin::SmallSpender => Ok(32),
+				pallet_origins::pallet::Origin::MediumSpender => Ok(33),
+				pallet_origins::pallet::Origin::BigSpender => Ok(34),
 				// DDC admins
-				origins::Origin::ClusterGovCreator => Ok(CLUSTER_ACTIVATOR_TRACK_ID),
-				origins::Origin::ClusterGovEditor => Ok(CLUSTER_ECONOMICS_UPDATER_TRACK_ID),
+				pallet_origins::pallet::Origin::ClusterGovCreator => Ok(CLUSTER_ACTIVATOR_TRACK_ID),
+				pallet_origins::pallet::Origin::ClusterGovEditor => {
+					Ok(CLUSTER_ECONOMICS_UPDATER_TRACK_ID)
+				},
 			}
 		} else {
 			Err(())
 		}
 	}
 }
+
 pallet_referenda::impl_tracksinfo_get!(TracksInfo, Balance, BlockNumber);
