+++ runtime/common/Cargo.toml
@@ -17,12 +17,14 @@ frame-system = { workspace = true }
 node-primitives = { workspace = true }
 pallet-contracts = { workspace = true }
 pallet-session = { workspace = true }
+pallet-referenda = { workspace = true }
 sp-api = { workspace = true }
 sp-core = { workspace = true }
 sp-io = { workspace = true }
 sp-runtime = { workspace = true }
 sp-staking = { workspace = true }
 sp-std = { workspace = true }
+sp-arithmetic = { workspace = true }
 
 [features]
 default = ["std"]
